monitor_service/requirements.txt
requests==2.31.0
mysql-connector-python==8.0.33
prometheus-client==0.17.1
schedule==1.2.0
